added FloatingPointWarningPass to xformer
