+++ tflite2xcore/tflite2xcore/transformation_passes/__init__.py
@@ -95,3 +95,5 @@ from .lce_passes import (
     ReplaceLceQuantizePass,
     LegalizeXCBconv2DPaddingPass,
 )
+
+from .warning_passes import FloatingPointWarningPass
